requirements.txt
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.0.1